M2/S2/LLMCP/TP2/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chess-challenge"
version = "0.1.0"
description = "LLM Chess Challenge - Train a 1M parameter model to play chess"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Nathanaël Fijalkow", email = "nathanael.fijalkow@gmail.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "accelerate>=0.26.0",
    "datasets>=2.14.0",
    "python-chess>=1.999",
    "huggingface-hub>=0.20.0",
    "tqdm>=4.65.0",
    "numpy>=1.24.0",
    "wandb>=0.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
eval = [
    "stockfish>=3.28.0",
]

[project.scripts]
chess-train = "src.train:main"
chess-eval = "src.evaluate:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
